+++ htfs/library.go
@@ -42,6 +42,7 @@ type Library interface {
 	Record([]byte) error
 	Restore([]byte, []byte, []byte) (string, error)
 	Location(string) string
+	ExactLocation(string) string
 	HasBlueprint([]byte) bool
 }
 
@@ -52,6 +53,10 @@ type hololib struct {
 
 func (it *hololib) Location(digest string) string {
 	return filepath.Join(common.HololibLocation(), "library", digest[:2], digest[2:4], digest[4:6])
+}
+
+func (it *hololib) ExactLocation(digest string) string {
+	return filepath.Join(common.HololibLocation(), "library", digest[:2], digest[2:4], digest[4:6], digest)
 }
 
 func (it *hololib) Identity() string {
@@ -196,6 +201,9 @@ func textual(key uint64, size int) string {
 }
 
 func makedirs(prefix string, suffixes ...string) error {
+	if common.Liveonly {
+		return nil
+	}
 	for _, suffix := range suffixes {
 		fullpath := filepath.Join(prefix, suffix)
 		err := os.MkdirAll(fullpath, 0o755)
@@ -216,4 +224,110 @@ func New() (Library, error) {
 		identity: sipit([]byte(basedir)),
 		basedir:  basedir,
 	}, nil
+}
+
+type virtual struct {
+	identity uint64
+	root     *Root
+	registry map[string]string
+	key      string
+}
+
+func Virtual() Library {
+	return &virtual{
+		identity: sipit([]byte(common.RobocorpHome())),
+	}
+}
+
+func (it *virtual) Identity() string {
+	return fmt.Sprintf("v%016xh", it.identity)
+}
+
+func (it *virtual) Stage() string {
+	stage := filepath.Join(common.HolotreeLocation(), it.Identity())
+	err := os.MkdirAll(stage, 0o755)
+	if err != nil {
+		panic(err)
+	}
+	return stage
+}
+
+func (it *virtual) Record(blueprint []byte) error {
+	defer common.Stopwatch("Holotree recording took:").Debug()
+	key := textual(sipit(blueprint), 0)
+	common.Timeline("holotree record start %s (virtual)", key)
+	fs, err := NewRoot(it.Stage())
+	if err != nil {
+		return err
+	}
+	err = fs.Lift()
+	if err != nil {
+		return err
+	}
+	common.Timeline("holotree (re)locator start (virtual)")
+	fs.AllFiles(Locator(it.Identity()))
+	common.Timeline("holotree (re)locator done (virtual)")
+	it.registry = make(map[string]string)
+	fs.Treetop(DigestMapper(it.registry))
+	fs.Blueprint = key
+	it.root = fs
+	it.key = key
+	return nil
+}
+
+func (it *virtual) Restore(blueprint, client, tag []byte) (string, error) {
+	defer common.Stopwatch("Holotree restore took:").Debug()
+	key := textual(sipit(blueprint), 0)
+	common.Timeline("holotree restore start %s (virtual)", key)
+	prefix := textual(sipit(client), 9)
+	suffix := textual(sipit(tag), 8)
+	name := prefix + "_" + suffix
+	metafile := filepath.Join(common.HolotreeLocation(), fmt.Sprintf("%s.meta", name))
+	targetdir := filepath.Join(common.HolotreeLocation(), name)
+	currentstate := make(map[string]string)
+	shadow, err := NewRoot(targetdir)
+	if err == nil {
+		err = shadow.LoadFrom(metafile)
+	}
+	if err == nil {
+		common.Timeline("holotree digest start (virtual)")
+		shadow.Treetop(DigestRecorder(currentstate))
+		common.Timeline("holotree digest done (virtual)")
+	}
+	fs := it.root
+	err = fs.Relocate(targetdir)
+	if err != nil {
+		return "", err
+	}
+	common.Timeline("holotree make branches start (virtual)")
+	err = fs.Treetop(MakeBranches)
+	common.Timeline("holotree make branches done (virtual)")
+	if err != nil {
+		return "", err
+	}
+	score := &stats{}
+	common.Timeline("holotree restore start (virtual)")
+	fs.AllDirs(RestoreDirectory(it, fs, currentstate, score))
+	common.Timeline("holotree restore done (virtual)")
+	defer common.Timeline("- dirty %d/%d", score.dirty, score.total)
+	common.Debug("Holotree dirty workload: %d/%d\n", score.dirty, score.total)
+	fs.Controller = string(client)
+	fs.Space = string(tag)
+	err = fs.SaveAs(metafile)
+	if err != nil {
+		return "", err
+	}
+	return targetdir, nil
+}
+
+func (it *virtual) ExactLocation(key string) string {
+	return it.registry[key]
+}
+
+func (it *virtual) Location(key string) string {
+	panic("Location is not supported on virtual holotree.")
+}
+
+func (it *virtual) HasBlueprint(blueprint []byte) bool {
+	return it.key == textual(sipit(blueprint), 0)
 }